glycin image backend
replace the Gdk::Texture image load/cache backend with a glycin based
on.

The raw Gdk::Texture backend would use 5x or more memory than the Gdk::Pixbuf
backend. The glycin backend still uses 2x-3x the memory, not great but still
better. This is also weird since glycin images get converted into a Gdk::Texture
anyway. Loading images with glycing also fixes the bad looking Gdk::Texture image
scaling.

+++ src/gui/main-window.cxx
@@ -1029,7 +1029,8 @@ gui::main_window::_draw_pages() noexcept
     std::vector<std::array<std::int32_t, 2>> size_list;
     for (const auto& pixbuf : pixbuf_list)
     {
-        size_list.push_back({pixbuf->get_width(), pixbuf->get_height()});
+        size_list.push_back({static_cast<std::int32_t>(pixbuf->get_width()),
+                             static_cast<std::int32_t>(pixbuf->get_height())});
     }
 
     // Rotation handling
